+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `L3_PACKET_MAX_SIZE` to `L3_CYPHERTEXT_MAX_SIZE`.
 - Renamed `L3_FRAME_MAX_SIZE` to `L3_PACKET_MAX_SIZE`.
 - Set `L3_RES_SIZE_SIZE` directly to value 2.
-
+- Renamed `STATIC` to `LT_STATIC` due to naming conflicts.
 
 ### Added
 - Macro `LT_CONFIG_OBJ_CNT` for number of objects in the configuration structure.